backend/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
asyncpg==0.31.0
pydantic==2.12.5
python-dotenv==1.0.1
